# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APP API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/bookmark": {
            "get": {
                "summary": "Get all user's bookmarks",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BookmarkController::index",
                "responses": {
                    "200": {
                        "description": "Bookmarks Listing"
                    }
                }
            },
            "put": {
                "summary": "Updates a business bio",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessBioController::update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Business ID",
                                        "type": "string"
                                    },
                                    "bio": {
                                        "description": "Business bio",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "summary": "Toggle bookmark specified by ID for logged in user",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BookmarkController::toggle",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bookmark successfully created/deleted!"
                    }
                }
            }
        },
...
        "/api/v1/business-posts/business/{$businessUuid}": {
            "get": {
                "summary": "Get all business posts for business",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessPostsController::forBusiness",
                "parameters": [
                    {
                        "name": "business_id",
                        "in": "query",
                        "description": "ID of business",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Business Post"
                    }
                }
            }
        },
...
        "/api/v1/businesses/stats": {
            "get": {
                "summary": "Get business stats",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessesController::stats",
                "parameters": [
                    {
                        "name": "top_left",
                        "in": "query",
                        "description": "Top Left of location (GPS)",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "bottom_right",
                        "in": "query",
                        "description": "Bottom right of location (GPS)",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stats data"
                    }
                }
            }
        },
...
        "/api/v1/businesses": {
            "post": {
                "summary": "Create a business",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessesController::store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name",
                                        "type": "string"
                                    },
                                    "lat": {
                                        "description": "latitude",
                                        "type": "string"
                                    },
                                    "lng": {
                                        "description": "longitude",
                                        "type": "string"
                                    },
                                    "bio": {
                                        "description": "business bio",
                                        "type": "string"
                                    },
                                    "avatar": {
                                        "description": "Avatar Image File",
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "description": "Category UUID",
                                        "type": "string"
                                    },
                                    "cover_photo": {
                                        "description": "Cover Photo Image File",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "BusinessResource"
                    }
                }
            }
        },
        "/api/v1/businesses/{business}": {
            "put": {
                "summary": "Update a business based on UUID passed.",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessesController::update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "base 64 encoded avatar inage",
                                        "type": "string"
                                    },
                                    "lat": {
                                        "description": "latitude",
                                        "type": "string"
                                    },
                                    "lng": {
                                        "description": "longitude",
                                        "type": "string"
                                    },
                                    "id": {
                                        "description": "business uuid",
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "description": "Category Id",
                                        "type": "integer"
                                    },
                                    "bio": {
                                        "description": "business bio",
                                        "type": "string"
                                    },
                                    "avatar": {
                                        "description": "business avatar",
                                        "type": "file"
                                    },
                                    "cover_photo": {
                                        "description": "business cover photo",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business updated"
                    },
                    "400": {
                        "description": "Business not found"
                    }
                }
            }
        },
        "/api/v1/businesses/{businessId}": {
            "delete": {
                "summary": "Delete a business by ID",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessesController::delete",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "business uuid",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business updated"
                    },
                    "400": {
                        "description": "Business not found"
                    }
                }
            }
        },
...
    }
}